[FIXED] Exceptions not exported
