+++ pyzeebe/__init__.py
@@ -1,5 +1,5 @@
+from pyzeebe import exceptions
 from pyzeebe.client.client import ZeebeClient
-from pyzeebe.common import exceptions
 from pyzeebe.credentials.camunda_cloud_credentials import CamundaCloudCredentials
 from pyzeebe.credentials.oauth_credentials import OAuthCredentials
 from pyzeebe.job.job import Job
