+++ pyzeebe/exceptions/message_exceptions.py
@@ -1,0 +1,5 @@
+from pyzeebe.exceptions.pyzeebe_exceptions import PyZeebeException
+
+
+class MessageAlreadyExists(PyZeebeException):
+    pass
