+++ pyzeebe/grpc_internals/zeebe_adapter_base.py
@@ -3,8 +3,8 @@ import os
 
 import grpc
 
-from pyzeebe.common.exceptions import ZeebeInternalError, ZeebeBackPressure, ZeebeGatewayUnavailable
 from pyzeebe.credentials.base_credentials import BaseCredentials
+from pyzeebe.exceptions.zeebe_exceptions import ZeebeBackPressure, ZeebeGatewayUnavailable, ZeebeInternalError
 from pyzeebe.grpc_internals.zeebe_pb2_grpc import GatewayStub
