+++ pyzeebe/grpc_internals/zeebe_job_adapter.py
@@ -4,7 +4,7 @@ from typing import Dict, List, Generator
 
 import grpc
 
-from pyzeebe.common.exceptions import JobNotFound, JobAlreadyDeactivated, ActivateJobsRequestInvalid
+from pyzeebe.exceptions import ActivateJobsRequestInvalid, JobAlreadyDeactivated, JobNotFound
 from pyzeebe.grpc_internals.zeebe_adapter_base import ZeebeAdapterBase
 from pyzeebe.grpc_internals.zeebe_pb2 import (ActivateJobsRequest, CompleteJobRequest,
                                               CompleteJobResponse, FailJobRequest, FailJobResponse, ThrowErrorRequest,
