+++ tests/integration/integration_test.py
@@ -5,7 +5,8 @@ from uuid import uuid4
 
 import pytest
 
-from pyzeebe import ZeebeWorker, ZeebeClient, exceptions, Job
+import pyzeebe.exceptions.workflow_exceptions
+from pyzeebe import ZeebeWorker, ZeebeClient, Job
 
 zeebe_client: ZeebeClient
 zeebe_worker = ZeebeWorker()
@@ -48,7 +49,7 @@ def test_run_workflow():
 
 
 def test_non_existent_workflow():
-    with pytest.raises(exceptions.WorkflowNotFound):
+    with pytest.raises(pyzeebe.exceptions.workflow_exceptions.WorkflowNotFound):
         zeebe_client.run_workflow(str(uuid4()))
