+++ tests/unit/worker/task_handler_test.py
@@ -3,7 +3,8 @@ from uuid import uuid4
 
 import pytest
 
-from pyzeebe.common.exceptions import TaskNotFound, NoVariableNameGiven
+from pyzeebe.exceptions import NoVariableNameGiven
+from pyzeebe.exceptions.pyzeebe_exceptions import TaskNotFound
 from pyzeebe.task.task import Task
 from pyzeebe.worker.task_handler import ZeebeTaskHandler, default_exception_handler
 from tests.unit.utils.random_utils import randint, random_job
